remove uneeded interfaces
