+++ components/function-runtimes/java11/serverless-java-sdk/src/main/java/io/project/kyma/serverless/sdk/Function.java
@@ -5,5 +5,5 @@ import javax.ws.rs.core.Response;
 
 
 public interface Function {
-    Response call(CloudEventImpl event, Context context);
+    Response call(CloudEvent event, Context context);
 }
