--- components/function-runtimes/java11/serverless-java-sdk/src/main/java/io/project/kyma/serverless/sdk/TraceableRequest.java
@@ -1,8 +1,0 @@
-package io.project.kyma.serverless.sdk;
-
-import javax.ws.rs.client.Invocation;
-
-public interface TraceableRequest {
-    Invocation.Builder getTraceableRequestBuilder(String target);
-
-}
